+++ requake/plot/plot_slip.py
@@ -12,10 +12,10 @@ Plot cumulative slip for one or more families.
 import logging
 import matplotlib as mpl
 import matplotlib.pyplot as plt
-import matplotlib.dates as mdates
 from matplotlib import cm
 from matplotlib import colors
 import numpy as np
+from .plot_utils import format_time_axis
 from ..families.families import FamilyNotFoundError, read_selected_families
 from ..formulas.slip import mag_to_slip_in_cm
 from ..config.rq_setup import rq_exit
@@ -37,13 +37,6 @@ def plot_slip(config):
         logger.error(m)
         rq_exit(1)
     fig, ax = plt.subplots(figsize=(8, 4))
-    years = mdates.YearLocator()   # every year
-    months = mdates.MonthLocator()  # every month
-    yearsFmt = mdates.DateFormatter('%Y')
-    ax.xaxis.set_major_locator(years)
-    ax.xaxis.set_major_formatter(yearsFmt)
-    ax.xaxis.set_minor_locator(months)
-    ax.xaxis.grid(True)
     ax.tick_params(which='both', top=True, labeltop=True)
     ax.tick_params(axis='x', which='both', direction='in')
 
@@ -64,6 +57,8 @@ def plot_slip(config):
             lw=1, marker='o', color=cmap(norm(fn % 10)),
             label=label)
         lines.append(line)
+    # time axis formatting must be done here, before setting limits below
+    format_time_axis(ax, which='xaxis')
     # get limits, that we will re-apply later
     xlim = ax.get_xlim()
     ylim = ax.get_ylim()