+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Copyright (c) 2021-2024 Claudio Satriano <satriano@ipgp.fr>
   read station metadata and waveform data from files. Supports any metadata
   format supported by ObsPy and SDS (SeisComp Data Structure) waveform
   archives.
+- Improved time axes in `plot_timespans` and `plot_slip` for short time
+  intervals
 
 ## v0.4.1 - 2024-03-11